server: split cursor test
